Added music assets
Added a set of music assets to the game.  At the moment there is only one song (A Million Dreams), but I am working on some more. It took a lot longer than I anticipated to get this made, but here it is. There is one track for each of our five instruments, with one reference track just for us.